successful test of create vm subcommand
