+++ cmd/auth.go
@@ -28,14 +28,7 @@ Required settings will be prompted based on provider`,
 func init() {
 	rootCmd.AddCommand(authCmd)
 
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// authCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// authCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+	// local flags
 	authCmd.Flags().StringP("provider", "p", "", "sets the cloud provider")
+	authCmd.MarkFlagRequired("provider")
 }
