+++ cmd/root.go
@@ -43,6 +43,8 @@ func init() {
 	// will be global for your application.
 
 	rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.maker.yaml)")
+	rootCmd.PersistentFlags().StringP("provider", "p", "", "sets the cloud provider")
+	rootCmd.MarkPersistentFlagRequired("provider")
 
 	// Cobra also supports local flags, which will only run
 	// when this action is called directly.
